Wire ConnectionRegistry into daemon startup (closes #9)
Replaces the single-client LLM path in main.rs with a registry build. The
daemon now constructs a ConnectionRegistry at startup, logs per-connection
availability, and takes the active connection's client out for the
ConversationHandler (still wrapped in retry / profiling layers). The
`[backend_tasks.llm]` path is left intact — #10 will reshape it into purpose
configs. If no connection builds successfully, the daemon falls back to a
client synthesized from the legacy `[llm]` block so existing installs keep
running until their config is migrated; the fallback emits a warn-level log
so the condition is visible.
